+++ src/main/java/com/backend/xplaza/model/Coupon.java
@@ -19,22 +19,22 @@ public class Coupon {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="coupon_id")
-    private long id;
+    private Long id;
 
     @Column(name="coupon_code")
     private String coupon_code;
 
     @Column(name="coupon_amount")
-    private float amount;
+    private Double amount;
 
     @Column(name="max_coupon_amount")
-    private float max_amount;
+    private Double max_amount;
 
     @Column(name="fk_currency_id")
-    private long currency_id;
+    private Long currency_id;
 
     @Column(name="fk_discount_type_id")
-    private long discount_type_id;
+    private Long discount_type_id;
 
     @Column(name="coupon_start_date")
     private Date start_date;
